+++ bc/recruitment/models.py
@@ -15,6 +15,7 @@ from django.shortcuts import get_object_or_404, render
 from django.template.loader import render_to_string
 from django.urls import reverse
 from django.utils.functional import cached_property
+from django.utils.safestring import mark_safe
 from django.utils.text import slugify
 
 from wagtail.admin.edit_handlers import FieldPanel, MultiFieldPanel, StreamFieldPanel
@@ -31,6 +32,7 @@ from ..utils.constants import RICH_TEXT_FEATURES
 from ..utils.email import NotifyEmailMessage
 from ..utils.models import BasePage
 from .constants import JOB_BOARD_CHOICES
+from .text_utils import extract_salary_range
 
 
 class JobSubcategory(models.Model):
@@ -197,7 +199,12 @@ class TalentLinkJob(models.Model):
     searchable_salary = models.CharField(
         max_length=255, help_text="Salary group for filtering"
     )
-    location = models.CharField(max_length=255)
+    location_name = models.CharField(max_length=255)
+    location_street_number = models.CharField(max_length=32, blank=True)
+    location_street = models.CharField(max_length=255, blank=True)
+    location_city = models.CharField(max_length=32, blank=True)
+    location_region = models.CharField(max_length=32, blank=True)
+    location_country = models.CharField(max_length=32, blank=True)
     location_postcode = models.CharField(max_length=8, blank=True)
     location_lat = models.DecimalField(
         max_digits=9, decimal_places=6, null=True, blank=True
@@ -216,6 +223,7 @@ class TalentLinkJob(models.Model):
         "images.CustomImage", null=True, related_name="+", on_delete=models.SET_NULL,
     )
     application_url_query = models.CharField(max_length=255)
+    organisation = models.TextField(blank=True)
 
     def get_categories_list(self):
         if self.subcategory:
@@ -245,6 +253,67 @@ class TalentLinkJob(models.Model):
             self.homepage = RecruitmentHomePage.objects.live().first()
 
         super().full_clean(*args, **kwargs)
+
+    @property
+    def schema_org_markup(self):
+        markup = {
+            "@context": "http://schema.org",
+            "@type": "JobPosting",
+            "datePosted": self.posting_start_date.isoformat(),
+            "description": self.short_description,
+            "title": self.title,
+            "validThrough": self.closing_date.isoformat(),
+            "employmentType": self.working_hours,
+            "hiringOrganization": {"@type": "Organization", "name": self.organisation},
+            "identifier": {
+                "@type": "PropertyValue",
+                "name": "Reference Number",
+                "value": self.job_number,
+            },
+            "jobLocation": {
+                "@type": "Place",
+                "name": self.location_name,
+                "address": {
+                    "@type": "PostalAddress",
+                    "streetAddress": " ".join(
+                        [self.location_street_number, self.location_street]
+                    ),
+                    "addressLocality": self.location_city,
+                    "addressRegion": self.location_region,
+                    "postalCode": self.location_postcode,
+                    "addressCountry": self.location_country,
+                },
+            },
+        }
+
+        if self.logo:
+            markup["hiringOrganization"]["logo"] = self.logo.get_rendition(
+                "max-110x110"
+            ).url
+
+        # Try to parse the salary_range, otherwise the searchable_salary.
+        for salary in (self.salary_range, self.searchable_salary):
+            try:
+                min_salary, max_salary = extract_salary_range(salary)
+            except TypeError:
+                pass
+            else:
+                markup["baseSalary"] = {
+                    "@type": "MonetaryAmount",
+                    "currency": "GBP",
+                    "value": {"@type": "QuantitativeValue", "unitText": "YEAR"},
+                }
+                if min_salary:
+                    markup["baseSalary"]["value"]["minValue"] = min_salary
+                if max_salary:
+                    markup["baseSalary"]["value"]["maxValue"] = max_salary
+                break
+
+        if self.location_lat and self.location_lon:
+            markup["jobLocation"]["latitude"] = str(self.location_lat)
+            markup["jobLocation"]["longitude"] = str(self.location_lon)
+
+        return mark_safe(json.dumps(markup))
 
 
 @receiver(pre_delete, sender=TalentLinkJob)