+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add Wagtail admin report of pages missing SEO metadata, and those with unpublished changes
 - Accessibility 91 Add in missing fieldsets for radio input groups
+- Add schema.org markup to job detail page
 
 ## 26.0 (2020-09-22)